refactor: change package name become robocup_client
