+++ include/robocup_client/robocup_client.hpp
@@ -18,11 +18,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef WEBOTS_CLIENT__WEBOTS_CLIENT_HPP_
-#define WEBOTS_CLIENT__WEBOTS_CLIENT_HPP_
+#ifndef ROBOCUP_CLIENT__ROBOCUP_CLIENT_HPP_
+#define ROBOCUP_CLIENT__ROBOCUP_CLIENT_HPP_
 
 #include "./messages.pb.h"
 #include "./robot_client.hpp"
 #include "./message_handler.hpp"
 
-#endif  // WEBOTS_CLIENT__WEBOTS_CLIENT_HPP_
+#endif  // ROBOCUP_CLIENT__ROBOCUP_CLIENT_HPP_
