+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>webots_client</name>
+  <name>robocup_client</name>
   <version>0.1.0</version>
   <description>ROS 2 client project for Webots API</description>
   <maintainer email="nathanterbaik@gmail.com">Nathanael</maintainer>
